jimt/LABS-472 - busted newly discovered createLibraryCLosure circle
