+++ apps/src/code-studio/components/libraries/libraryParser.js
@@ -1,5 +1,5 @@
 import {getFunctionsAndMetadata} from '@cdo/apps/lib/tools/jsinterpreter/JSInterpreter';
-
+import {createLibraryClosure} from './libraryParser/createLibraryClosure';
 /**
  * Finds all functions in the provided code and returns them with any metadata
  * @param {string} code The code to parse to get all functions
@@ -49,28 +49,6 @@ function createDropletConfig(functions, libraryName) {
   }
 
   return fullConfig;
-}
-
-/**
- * Creates a closure to contain the library
- * The format will look like this:
- * var libraryName = (function() {function myFunc{doSomething}; return {myFunc: myFunc}})();
- * The library functions can be called with dot notation:
- * libraryName.myFunc();
- *
- * @param {object} json The library and all metadata
- *   json {
- *     functions: Array<string>
- *     source: string
- *     name: string
- *   }
- * @returns {string} closure The library closure
- */
-export function createLibraryClosure(json) {
-  let exportedFunctions = [];
-  exportedFunctions = json.functions.map(name => `${name}: ${name}`);
-  let functionsInClosure = exportedFunctions.join(',');
-  return `var ${json.name} = (function() {${json.source};\nreturn {${functionsInClosure}}})();`;
 }
 
 /**
